Make everything private
