+++ packages/flutter/lib/cupertino.dart
@@ -33,8 +33,6 @@ export 'src/cupertino/context_menu.dart';
 export 'src/cupertino/context_menu_action.dart';
 export 'src/cupertino/date_picker.dart';
 export 'src/cupertino/desktop_text_selection.dart';
-export 'src/cupertino/desktop_text_selection_toolbar.dart';
-export 'src/cupertino/desktop_text_selection_toolbar_button.dart';
 export 'src/cupertino/dialog.dart';
 export 'src/cupertino/form_row.dart';
 export 'src/cupertino/form_section.dart';
